Fix misuse of "int" type in lib/char_ref.rl

+++ nokogumbo-import/gumbo-parser/src/char_ref.c
@@ -20783,7 +20783,7 @@ _again:
   if (cs >= 7623) {
     assert(output->first != kGumboNoChar);
     char last_char = *(te - 1);
-    int len = te - start;
+    size_t len = te - start;
     if (last_char == ';') {
       bool matched = utf8iterator_maybe_consume_match(input, start, len, true);
       assert(matched);
